refactor: re-organize config type files
