+++ packages/analytics-browser/test/helpers/default.ts
@@ -1,10 +1,10 @@
 import { MemoryStorage } from '@amplitude/analytics-core';
-import { BrowserConfig as IBrowserConfig, InitOptions, UserSession } from '@amplitude/analytics-types';
+import { BrowserConfig as IBrowserConfig, BrowserOptions, UserSession } from '@amplitude/analytics-types';
 import { SessionManager } from '@amplitude/analytics-client-common';
 
 import { BrowserConfig } from '../../src/config';
 
-export const useDefaultConfig = (userId?: string, overrides?: Partial<InitOptions<BrowserConfig>>) =>
+export const useDefaultConfig = (userId?: string, overrides?: BrowserOptions) =>
   new BrowserConfig(API_KEY, userId || USER_ID, { ...DEFAULT_OPTIONS, ...overrides });
 
 export const API_KEY = 'apiKey';
@@ -13,7 +13,7 @@ export const USER_ID = 'userId';
 
 const cookieStorage = new MemoryStorage<UserSession>();
 
-export const DEFAULT_OPTIONS: InitOptions<IBrowserConfig> = {
+export const DEFAULT_OPTIONS: Partial<IBrowserConfig> = {
   apiKey: API_KEY,
   cookieStorage,
   cookieExpiration: 365,
