+++ packages/analytics-react-native/test/helpers/default.ts
@@ -1,10 +1,10 @@
 import { MemoryStorage } from '@amplitude/analytics-core';
-import { ReactNativeConfig as IReactNativeConfig, InitOptions, UserSession } from '@amplitude/analytics-types';
+import { ReactNativeConfig as IReactNativeConfig, ReactNativeOptions, UserSession } from '@amplitude/analytics-types';
 import { SessionManager } from '@amplitude/analytics-client-common';
 
 import { ReactNativeConfig } from '../../src/config';
 
-export const useDefaultConfig = (userId?: string, overrides?: Partial<InitOptions<ReactNativeConfig>>) =>
+export const useDefaultConfig = (userId?: string, overrides?: ReactNativeOptions) =>
   new ReactNativeConfig(API_KEY, userId || USER_ID, { ...DEFAULT_OPTIONS, ...overrides });
 
 export const API_KEY = 'apiKey';
@@ -13,7 +13,7 @@ export const USER_ID = 'userId';
 
 const cookieStorage = new MemoryStorage<UserSession>();
 
-export const DEFAULT_OPTIONS: InitOptions<IReactNativeConfig> = {
+export const DEFAULT_OPTIONS: Partial<IReactNativeConfig> = {
   apiKey: API_KEY,
   cookieStorage,
   cookieExpiration: 365,
