+++ packages/analytics-types/src/config/node.ts
@@ -1,0 +1,5 @@
+import { Config } from './core';
+
+export type NodeConfig = Config;
+
+export type NodeOptions = Omit<Partial<NodeConfig>, 'apiKey'>;
