+++ packages/analytics-types/src/config/react-native.ts
@@ -1,0 +1,16 @@
+import { TransportType } from '../transport';
+import { BrowserConfig, TrackingOptions } from './browser';
+
+export interface ReactNativeConfig extends BrowserConfig {
+  trackingOptions: ReactNativeTrackingOptions;
+  trackingSessionEvents?: boolean;
+}
+
+export interface ReactNativeTrackingOptions extends TrackingOptions {
+  adid?: boolean;
+  carrier?: boolean;
+}
+
+export interface ReactNativeOptions extends Omit<Partial<ReactNativeConfig>, 'apiKey'> {
+  transport?: TransportType | keyof typeof TransportType;
+}
