+++ packages/analytics-types/src/server-zone.ts
@@ -1,0 +1,4 @@
+export enum ServerZone {
+  US = 'US',
+  EU = 'EU',
+}
